Created a DialogBox class that automatically creates a DialogBox for you, you just have to give it a Rect to signify its location.  Also you can use setText to change its text.  So we'll most likely just need one instance of this.
Messed with the Player class's method addPokemon, I think this needs to
be changed to just add a single pokemon and not use a "while" so it
needs to fill up the array.

Added a constructor to RFIDListener and a boolean called "waiter" that
will wait while fetching an RFID tag.

Commented out the prototype code.

Created the initial DialogBox and will soon work on placing initial
pokemon on the board.